#931 - tidy up file planner code
